Add files via upload
Art gallery management is to manage the works of an art gallery. As simple as it might sound, the management of an art gallery is quite complex. It has several artworks from various artists, sometimes the same. They are displayed in museums and halls to showcase the artists’ talents. Customers who come in to admire these intrinsic artworks eventually purchase them too. All of this is managed systematically for the art gallery to function properly. Hence, the central purpose of our project is to create a database management system that allows the smooth sailing of these art galleries. This system intends to help customers access our artworks, exhibition/event dates, and the art galleries that exhibit these artworks.